# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GMRAG API",
    "description": "GMRAG 2.0 HTTP API. Authenticated routes require `Authorization: Bearer <JWT>`. Tenant-scoped routes under `/tenants/{tid}/...` also require header `X-Tenant-ID` matching the path `{tid}`. List endpoints return full result sets (no pagination) unless noted.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/tenants/{tid}/settings/llm": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Read tenant LLM / BYOK settings (owner-only, masked API key).",
        "operationId": "get_llm_settings",
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "description": "Must match path tid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "LLM settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — owner only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Authorization unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Upsert tenant LLM / BYOK settings (owner-only).",
        "operationId": "put_llm_settings",
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "description": "Must match path tid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutLlmSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated LLM settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — owner only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Authorization unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Machine-readable error code (kebab-case) and human-readable message.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response envelope.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "LlmSettingsResponse": {
        "type": "object",
        "required": [
          "configured",
          "has_api_key"
        ],
        "properties": {
          "api_key_masked": {
            "type": [
              "string",
              "null"
            ]
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "configured": {
            "type": "boolean"
          },
          "dimensions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "has_api_key": {
            "type": "boolean"
          },
          "llm_base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "llm_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PutLlmSettingsRequest": {
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "dimensions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "llm_base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "llm_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak OIDC access token. Use Authorize and paste the JWT (with or without the `Bearer ` prefix)."
      }
    }
  },
...
}